Add files via upload
Grade converter is a program that allows you to convert between 4 different metrics (Miles and Kilometers, Fahrenheit & Celsius, Feet & Meters, Pounds and Kilograms), for each options it will convert either way (ex: Miles to Kilometers or Kilometers to Miles). To use, simply follow the prompts and input the number with the corresponding option.